+++ Back-End/src/main/java/com/senac/aesthetics/domain/Servico.java
@@ -1,5 +1,6 @@
 package com.senac.aesthetics.domain;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 import jakarta.persistence.Column;
@@ -47,13 +48,13 @@ public class Servico {
     @Digits(integer = 10, fraction = 2, message = "O Valor Máximo do Preço de Custo do Serviço é de R$ 1.000.000,00 (1 Milhão)!")
     @PositiveOrZero(message = "O Valor do Preço de Custo do Serviço Deve Estar Entre R$ 0 (Zero) e R$ 1.000.000,00 (1 Milhão)!")
     @NotNull(message = "O Valor do Preço de Custo do Serviço Deve Ser Informado!")
-    private Float precoCusto;
+    private BigDecimal precoCusto;
     
     @Column(name = "PRECO_VENDA", nullable = false, precision = 10, scale = 2)
     @Digits(integer = 10, fraction = 2, message = "O Valor Máximo do Preço de Venda do Serviço é de R$ 1.000.000,00 (1 Milhão)!")
     @PositiveOrZero(message = "O Valor do Preço de Venda do Serviço Deve Estar Entre R$ 0 (Zero) e R$ 1.000.000,00 (1 Milhão)!")
     @NotNull(message = "O Valor do Preço de Venda do Serviço Deve Ser Informado!")
-    private Float precoVenda;
+    private BigDecimal precoVenda;
 
     // Relacionamentos:
     @OneToMany(orphanRemoval = false, mappedBy = "servico")
